refactor: merge ArrayTask and RunOrReplicate
- Have it always replicate / make array. i wanted the ui to be cleaner, but that should be a non-goal.

- Re-use ArrayTask as final name but base on RunOrReplicate
