+++ packages/task-graph/src/task/TaskError.ts
@@ -80,45 +80,3 @@ export class TaskInvalidInputError extends TaskError {
     super(message);
   }
 }
-
-/**
- * Compound tasks can have multiple errors, and this error is used to group them together.
- *
- * Examples: Task with isCompound=true, ArrayTask, TaskRunner, etc.
- */
-export class TaskErrorGroup extends TaskError {
-  static readonly type: string = "TaskErrorGroup";
-  constructor(private errors: { key: unknown; type: string; error: TaskError }[]) {
-    super("Child errors occurred");
-  }
-  getError(key: unknown): TaskError | undefined {
-    return this.errors.find((e) => e.key === key)?.error;
-  }
-  getErrors(): TaskError[] {
-    return this.errors.map((e) => e.error);
-  }
-  getErrorKeys(): unknown[] {
-    return this.errors.map((e) => e.key);
-  }
-  getErrorCount(): number {
-    return this.errors.length;
-  }
-  getErrorNames(): string[] {
-    return this.errors.map((e) => e.error.name);
-  }
-  getErrorMessages(): string[] {
-    return this.errors.map((e) => e.error.message);
-  }
-  getFirstError(): TaskError {
-    return this.errors[0].error;
-  }
-  getFirstErrorKey(): unknown {
-    return this.errors[0].key;
-  }
-  getFirstErrorType(): string {
-    return this.errors[0].type;
-  }
-  hasAbortError(): boolean {
-    return this.errors.some((e) => e.error instanceof TaskAbortedError);
-  }
-}
